Extract buildFieldsTree into a function and use to generate ThemeCode
