+++ TODO.md
@@ -13,7 +13,8 @@
 
 ## Merge
 [ ] Make code generator work again
-[ ] Check theme code
+[ ] Polylang: live update with list of fields
+[ ] Check theme code: live update with list of fields
 [ ] Fix translation is loaded by site language, which should be by user language
 
 ## Improvements
