+++ app/list-functions.js
@@ -204,6 +204,38 @@ const getList = id => {
 	console.error( `%cList with id "${ id }" not found.`, 'color:red' );
 };
 
+export const buildFieldsTree = () => {
+	// Get all root fields first
+	const rootFields = lists.has( 'root' )
+		? [ ...lists.get( 'root' ).getState().fields ]
+		: [];
+
+	// Process each field to include sub-fields for groups
+	const processField = field => {
+		// Deep clone the field to avoid reference issues
+		const processedField = { ...field };
+
+		// Temporary keys used in the builder.
+		delete processedField._new;
+		delete processedField._active;
+		delete processedField._id_changed;
+
+		// Temporary keys used by SortableJS.
+		delete processedField.chosen;
+		delete processedField.selected;
+
+		// If it's a group field, get its sub-fields
+		if ( field.type === 'group' && lists.has( field._id ) ) {
+			processedField.fields = lists.get( field._id ).getState().fields.map( processField );
+		}
+
+		return processedField;
+	};
+
+	// Process all root fields
+	return rootFields.map( processField );
+};
+
 // Parse fields and put into the lists.
 // Recursively put groups' fields into other lists.
 const parseLists = ( obj, listId ) => {
